+++ test/http.js
@@ -131,6 +131,46 @@ buster.testCase('http - request', {
       assert.isNull(err);
       done();
     });
+  },
+  'should set headers': function (done) {
+    this.stub(request, 'post', function (params, cb) {
+      assert.equals(params.headers.foo, 'bar');
+      cb(null, { statusCode: 200 });
+    });
+    function _success(result, cb) {
+      cb(null, result);
+    }
+    http.request('POST', 'http://someurl', { headers: { foo: 'bar' }, handlers: { 200: _success }}, function (err, result) {
+      assert.isNull(err);
+      done();
+    });
+  },
+  'should set request payload': function (done) {
+    this.stub(request, 'post', function (params, cb) {
+      assert.equals(params.json, '{ "foo": "bar" }');
+      cb(null, { statusCode: 200 });
+    });
+    function _success(result, cb) {
+      cb(null, result);
+    }
+    http.request('POST', 'http://someurl', { json: '{ "foo": "bar" }', handlers: { 200: _success }}, function (err, result) {
+      assert.isNull(err);
+      done();
+    });
+  },
+  'should override http#req params when requestOpts is provided': function (done) {
+    this.stub(process, 'env', { http_proxy: 'http://someproxy', https_proxy: 'https://someproxy' });
+    this.stub(request, 'get', function (params, cb) {
+      assert.equals(params.proxy, 'http://overrideproxy');
+      cb(null, { statusCode: 200, body: 'somebody' });
+    });
+    function _success(result, cb) {
+      cb(null, result);
+    }
+    http.request('GET', 'http://someurl', { requestOpts: { proxy: 'http://overrideproxy' }, handlers: { 200: _success }}, function (err, result) {
+      assert.isNull(err);
+      done();
+    });
   }
 });